+++ app/models/audio_event.rb
@@ -63,7 +63,7 @@ class AudioEvent < ActiveRecord::Base
                        :low_frequency_hertz, :high_frequency_hertz,
                        :is_reference,
                        :created_at, :creator_id, :updated_at],
-        render_fields: [:id, :audio_recording_id, :comment, :flag,
+        render_fields: [:id, :audio_recording_id,
                         :start_time_seconds, :end_time_seconds,
                         :low_frequency_hertz, :high_frequency_hertz,
                         :is_reference,